server/requirements.txt
asgiref==3.8.1
certifi==2024.2.2
charset-normalizer==3.3.2
Django==5.0.6
django-cors-headers==4.3.1
django-rest-framework==0.1.0
django-sendfile==0.3.11
djangorestframework==3.15.1
idna==3.7
requests==2.31.0
sqlparse==0.5.0
urllib3==2.2.1
gunicorn==22.0.0
psycopg2-binary==2.9.9
pyjwt==2.8.0